// prj23Bookings-HttpActions-Adding-Fetching-Deleting-FireBase/src/app/Utilities/Interfaces/bookingsdatafirebase.interface.ts
//This is a places holder data structure for fetching Place(s) from the Firbase
export interface IBookingDataFireBase{
  bookedFrom: string;
  bookedTo: string;
  firstName: string;
  lastName: string;
  guestNumber: number;
  placeId: string;
  placeImage: string;
  placeTitle: string;
  userId: string
}